requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
timm==0.9.0
lightning==2.1.3
ultralytics>=8.1.7
Pillow==10.0.0
requests=2.31.0
pandas==2.2.0
numpy==1.26.3
supervision==0.18.0
opencv-python-headless==4.8.1.78
boto3==1.34.34 
python-dotenv>=1.0.1
botocore==1.34.49 